Refs #14181 - Fix authentication validation when creating registry
On 9ea405911cef1f02ca42db77c16bea4fe297093e Docker.authenticate! was
calling /auth on the Docker host API. This kind of authentication was
not correct, as it was authenticating access to the Docker host, not to
the registry.

Instead, we just perform a restclient call with basic authentication
against an endpoint as per https://docs.docker.com/v1.7/reference/api/hub_registry_spec/#authentication-and-authorization

We should support token authentication as well after this and add a
new field to the form.